+++ wristband/apps/tests/test_providers.py
@@ -1,0 +1,4 @@
+import mock
+
+from wristband.apps.providers import DocktorAppDataProvider, NestedDocktorAppDataProvider, GenericDocktorDataProvider
+
