--- wristband/providers/tests/test_utils.py
@@ -1,37 +1,0 @@
-import pytest
-import mock
-
-from wristband.common.utils import extract_stage, extract_security_zone_from_env, get_security_zone_from_app_name
-
-@pytest.mark.parametrize(('environment', 'expected_result'), [
-    ('foo-bar', 'foo'),
-    ('dev', 'dev')
-])
-def test_extract_stage(environment, expected_result):
-    assert extract_stage(environment) == expected_result
-
-
-@pytest.mark.parametrize(('environment', 'expected_result'), [
-    ('foo-bar', 'bar'),
-    ('dev', 'dev')
-])
-def test_extract_security_zone_from_env(environment, expected_result):
-    assert extract_security_zone_from_env(environment) == expected_result
-
-
-def test_get_security_zone_from_app_name_app_found(dummy_app_class):
-    with mock.patch('wristband.common.utils.App') as mocked_app:
-        mocked_objects_method = mock.Mock()
-        mocked_objects_method.first.return_value = dummy_app_class(name='foo', security_zone='bar')
-        mocked_app.objects.return_value = mocked_objects_method
-
-        assert get_security_zone_from_app_name('foo') == 'bar'
-
-
-def test_get_security_zone_from_app_name_app_not_found(dummy_app_class):
-    with mock.patch('wristband.common.utils.App') as mocked_app:
-        mocked_objects_method = mock.Mock()
-        mocked_objects_method.first.return_value = []
-        mocked_app.objects.return_value = mocked_objects_method
-
-        assert get_security_zone_from_app_name('foo') is None
